% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ph.R
\name{is.ph}
\alias{is.ph}
\title{Create or test for dates of public holidays}
\usage{
is.ph(date, oil = T)
}
\arguments{
\item{date}{A date vector to be tested.}

\item{oil}{Logical. Default \code{TRUE} will include off-in-lieu.}
}
\value{
A logical vector indicating public holiday or off-in-lieu.
}
\description{
Create or test for dates of public holidays, with option for off-in-lieu.
}
\examples{
library(dplyr)
library(lubridate)
data.frame(cal = seq.Date(ymd(20130101), ymd(20130401), 1) ) \%>\% 
mutate(is_ph = is.ph(cal, TRUE))
}
