% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.workday.R
\name{is.workday}
\alias{is.workday}
\title{Test for working days}
\usage{
is.workday(date, oil = T, weekend = c("Sat", "Sun"))
}
\arguments{
\item{date}{A date vector to be tested.}

\item{oil}{Logical. Default \code{TRUE} will account for public holiday off-in-lieu.}

\item{weekend}{Vector of days to define weekend. Default include Sat and Sun.}
}
\value{
A logical vector indicating working day or not.
}
\description{
Test for dates of working days, with options to define weekends and
whether to exclude public holiday off-in-lieu.
}
\examples{
library(dplyr)
library(lubridate)
library(sgholidays)
data.frame(cal = seq.Date(ymd(20130101), ymd(20130401), 1) ) \%>\%
  dplyr::mutate(day = lubridate::wday(cal, TRUE)) \%>\% 
  mutate(is_ph = is.ph(cal, TRUE)) \%>\% 
  mutate(is_workday = is.workday(cal, TRUE))
}
