% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgholidays-package.R
\docType{package}
\name{sgholidays-package}
\alias{sgholidays}
\alias{sgholidays-package}
\title{sgholidays: Data and functions to deal with holidays in Singapore}
\description{
Built-in data of Singapore's gazetted public holidays since 2008. Functions to deal with holidays in Singapore are in development.
}
\author{
\strong{Maintainer}: Tse Lert Chua \email{tselert@hotmail.com} (\href{https://orcid.org/0000-0003-3416-1032}{ORCID})

}
\keyword{internal}
