% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgph.R
\docType{data}
\name{sgph}
\alias{sgph}
\title{Data of Public Holidays in Singapore}
\format{
A data.table with 2 variables:
\describe{
\item{PublicHolidayName}{chr Name of public holiday}
\item{Date}{Date Date of gazetted public holiday}
}
}
\source{
\url{https://www.mom.gov.sg/employment-practices/public-holidays 
https://www.mom.gov.sg/newsroom/press-releases/2007/singapore-public-holidays-for-the-year-2008 
https://sso.agc.gov.sg/Act/HA1998#Sc-
https://www.mom.gov.sg/employment-practices/public-holidays-entitlement-and-pay
}
}
\usage{
sgph
}
\description{
This dataset contains all the dates of gazetted public holidays (PH) in Singapore since 2008.
It will be updated annually as per the official website of the Ministry of Manpower Singapore.
}
\details{
Since Aug 1968, Singaporeans enjoy 11 gazetted public holidays:
\enumerate{
\item New Year Day (1d)
\item Chinese New Year (2d)
\item Good Friday (1d)
\item Labour Day (1d)
\item Hari Raya Puasa (1d)
\item Vesak Day (1d)
\item National Day (1d)
\item Hari Raya Haji (1d)
\item Deepavali (1d)
\item Christmas Day (1d)
}

In addition, Polling Day is also a public holiday for Presidential Election and General Elections

Fun fact: Singapore scrapped a few holidays in August 1968 to improve productivity.

Note on Off-in-lieu (OIL):
This dataset does not contain dates of OIL. Function is in development.

If the gazetted PH falls on a Sunday, the following working day will be the 'OIL' by default.

If the gazetted PH falls on a worker's rest day (especially Saturday), he may be granted an 'OIL' on a date agreed with his/her employer or be
compensated monetarily.
}
\keyword{datasets}
